going full float for all positions

+++ src/Game.hpp
@@ -120,7 +120,7 @@ class Game : public Engine {
 		pathfindingSystem = std::make_unique<PathfindingSystem>(mapManager);
 	}
 
-	void createTestEntity(const Vec2i &position, const std::vector<PatrolPoint> &waypoints)
+	void createTestEntity(const Vec2f &position, const std::vector<PatrolPoint> &waypoints)
 	{
 		Entity e = instantiateNPCEntity(ecs, position);
 		ecs.addComponent<Vision>(e, Vision{});
@@ -131,17 +131,17 @@ class Game : public Engine {
 
 	void addTestEntities()
 	{
-		createTestEntity({15, 6}, {{Vec2i{15, 6} * TILE_SIZE, Rotation::SOUTH, 2},
-		                           {Vec2i{19, 5} * TILE_SIZE, Rotation::NORTH, 2},
-		                           {Vec2i{24, 7} * TILE_SIZE, Rotation::EAST, 2}});
+		createTestEntity({15, 6}, {{Vec2f{15, 6} * TILE_SIZE, Rotation::SOUTH, 2},
+		                           {Vec2f{19, 5} * TILE_SIZE, Rotation::NORTH, 2},
+		                           {Vec2f{24, 7} * TILE_SIZE, Rotation::EAST, 2}});
 		createTestEntity(
-		    {21, 7}, {{Vec2i{21, 7} * TILE_SIZE, Rotation::SOUTH, 2}, {Vec2i{21, 8} * TILE_SIZE, Rotation::NORTH, 2}});
+		    {21, 7}, {{Vec2f{21, 7} * TILE_SIZE, Rotation::SOUTH, 2}, {Vec2f{21, 8} * TILE_SIZE, Rotation::NORTH, 2}});
 		createTestEntity(
-		    {14, 7}, {{Vec2i{14, 7} * TILE_SIZE, Rotation::EAST, 5}, {Vec2i{20, 7} * TILE_SIZE, Rotation::WEST, 5}});
-		createTestEntity({16, 6}, {{Vec2i{16, 6} * TILE_SIZE, Rotation::SOUTH, 0},
-		                           {Vec2i{19, 6} * TILE_SIZE, Rotation::NORTH, 0},
-		                           {Vec2i{19, 7} * TILE_SIZE, Rotation::EAST, 0},
-		                           {Vec2i{16, 7} * TILE_SIZE, Rotation::WEST, 0}});
+		    {14, 7}, {{Vec2f{14, 7} * TILE_SIZE, Rotation::EAST, 5}, {Vec2f{20, 7} * TILE_SIZE, Rotation::WEST, 5}});
+		createTestEntity({16, 6}, {{Vec2f{16, 6} * TILE_SIZE, Rotation::SOUTH, 0},
+		                           {Vec2f{19, 6} * TILE_SIZE, Rotation::NORTH, 0},
+		                           {Vec2f{19, 7} * TILE_SIZE, Rotation::EAST, 0},
+		                           {Vec2f{16, 7} * TILE_SIZE, Rotation::WEST, 0}});
 	}
 
 	ECSManager ecs;
