+++ src/ai/nodes/MoveTo.hpp
@@ -4,7 +4,7 @@
 #include "../../components/Positionable.hpp"
 #include "../../components/RigidBody.hpp"
 #include "../../ecs/ECSManager.hpp"
-#include "../../engine/types/Vec2i.hpp"
+#include "../../engine/types/Vec2f.hpp"
 #include "../../modules/AStar.hpp"
 #include "behaviortree_cpp/action_node.h"
 #include "behaviortree_cpp/basic_types.h" // ports etc
@@ -24,7 +24,7 @@ class MoveTo : public BT::StatefulActionNode {
 		// clang-format off
 		return {
 			BT::InputPort<Entity>("entity"),
-			BT::InputPort<Vec2i>("position")
+			BT::InputPort<Vec2f>("position")
 		};
 		// clang-format on
 	}
@@ -38,13 +38,13 @@ class MoveTo : public BT::StatefulActionNode {
 	BT::NodeStatus onRunning() override
 	{
 		BT::Expected<Entity> entity = getInput<Entity>("entity");
-		BT::Expected<Vec2i> pos = getInput<Vec2i>("position");
+		BT::Expected<Vec2f> pos = getInput<Vec2f>("position");
 
 		if (!entity)
 			return BT::NodeStatus::FAILURE;
 
 		auto &ai = ecs.getComponent<AI>(entity.value());
-		const Vec2i &position = ecs.getComponent<Positionable>(entity.value()).position;
+		const Vec2f &position = ecs.getComponent<Positionable>(entity.value()).position;
 		ai.targetPosition = pos.value();
 
 		if (position == ai.targetPosition)
