+++ src/entities/sign.hpp
@@ -13,7 +13,8 @@ Entity instantiateSignEntity(ECSManager &ecs, Vec2i positionInTiles, const std::
 {
 	Entity npc = ecs.addEntity();
 
-	ecs.addComponent(npc, Positionable{{positionInTiles.x * TILE_SIZE, positionInTiles.y * TILE_SIZE}});
+	ecs.addComponent(npc,
+	                 Positionable{{(float)(positionInTiles.x * TILE_SIZE), (float)(positionInTiles.y * TILE_SIZE)}});
 	ecs.addComponent(npc, Renderable{{5 * TILE_SIZE, 1 * TILE_SIZE}, {TILE_SIZE, TILE_SIZE}, 0});
 	ecs.addComponent(npc, Collider{});
 	ecs.addComponent(npc, Interactable{text});
